AppLauncher Button & Finalized OptionWindow
- Added ref for Assembly-CSharp-firstpass.dll
- Added image for AppLauncher & header image for the OptionsWindow

#482
